Initial Commit
This is a simple maths quiz app being made for my Treehouse Techdegree.
At the moment, the app generates an addition problem and provides the
user with three possible answers. When user submits their answer, the
app then produces a toast which congratulates the user if they got the
answer correct and advises the correct answer if they got the answer
wrong.

Still to come: improved styling, adding subtraction, multiplication &
division questions to the app.